+++ src/main/java/de/aivot/GoverBackend/services/pdf/NumberFormatDialect.java
@@ -1,5 +1,6 @@
 package de.aivot.GoverBackend.services.pdf;
 
+import jakarta.annotation.Nullable;
 import org.thymeleaf.context.IExpressionContext;
 import org.thymeleaf.dialect.AbstractDialect;
 import org.thymeleaf.dialect.IExpressionObjectDialect;
@@ -21,12 +22,22 @@ public class NumberFormatDialect extends AbstractDialect implements IExpressionO
         super("numberformat");
     }
 
-    public String format(Object value, int decimalPlaces) {
-        Locale locale = Locale.GERMAN;
-        NumberFormat formatter = NumberFormat.getNumberInstance(locale);
-        formatter.setMinimumFractionDigits(decimalPlaces);
-        formatter.setMaximumFractionDigits(decimalPlaces);
-        return formatter.format(value);
+    public String format(@Nullable Object value, @Nullable Integer decimalPlaces) {
+        if (value == null) {
+            return "";
+        }
+
+        if (value instanceof Number nValue) {
+            Locale locale = Locale.GERMAN;
+            NumberFormat formatter = NumberFormat.getNumberInstance(locale);
+            formatter.setMinimumFractionDigits(decimalPlaces == null ? 0 : decimalPlaces);
+            formatter.setMaximumFractionDigits(decimalPlaces == null ? 0 : decimalPlaces);
+            return formatter.format(nValue);
+        } else if (value instanceof String s) {
+            return s;
+        } else {
+            return "";
+        }
     }
 
     public String formatISOTimestamp(String timestamp, String format) {